basic-RAG/pyproject.toml
[tool.poetry]
name = "basic-rag"
version = "0.1.0"
description = ""
authors = ["Anantvir"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.14"
langchain-openai = "^0.1.22"
langchain-community = "^0.2.12"
python-dotenv = "^1.0.1"
bs4 = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
